backend/requirements.txt
PyPDF2
PyMuPDF
pytesseract
Pillow
flask
flask_cors
anthropic
gunicorn
